done through ScoringCategoryService

+++ Cite.Api/Controllers/MoveController.cs
@@ -6,8 +6,9 @@ using System.Collections.Generic;
 using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Cite.Api.Data.Enumerations;
+using Cite.Api.Infrastructure.Authorization;
 using Cite.Api.Infrastructure.Extensions;
 using Cite.Api.Infrastructure.Exceptions;
 using Cite.Api.Infrastructure.QueryParameters;
@@ -20,9 +21,9 @@ namespace Cite.Api.Controllers
     public class MoveController : BaseController
     {
         private readonly IMoveService _moveService;
-        private readonly IAuthorizationService _authorizationService;
+        private readonly ICiteAuthorizationService _authorizationService;
 
-        public MoveController(IMoveService moveService, IAuthorizationService authorizationService)
+        public MoveController(IMoveService moveService, ICiteAuthorizationService authorizationService)
         {
             _moveService = moveService;
             _authorizationService = authorizationService;
@@ -42,6 +43,9 @@ namespace Cite.Api.Controllers
         [SwaggerOperation(OperationId = "getByEvaluation")]
         public async Task<IActionResult> GetByEvaluation(Guid evaluationId, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<Evaluation>(evaluationId, [SystemPermission.ViewEvaluations, SystemPermission.ObserveEvaluations], [EvaluationPermission.ObserveEvaluation, EvaluationPermission.ViewEvaluation], ct))
+                throw new ForbiddenException();
+
             var list = await _moveService.GetByEvaluationAsync(evaluationId, ct);
             return Ok(list);
         }
@@ -60,8 +64,10 @@ namespace Cite.Api.Controllers
         [SwaggerOperation(OperationId = "getMove")]
         public async Task<IActionResult> Get(Guid id, CancellationToken ct)
         {
-            var move = await _moveService.GetAsync(id, ct);
+            if (!await _authorizationService.AuthorizeAsync<Move>(id, [SystemPermission.ViewEvaluations, SystemPermission.ObserveEvaluations], [EvaluationPermission.ObserveEvaluation, EvaluationPermission.ViewEvaluation], ct))
+                throw new ForbiddenException();
 
+            var move = await _moveService.GetAsync(id, ct);
             if (move == null)
                 throw new EntityNotFoundException<Move>();
 
@@ -83,6 +89,9 @@ namespace Cite.Api.Controllers
         [SwaggerOperation(OperationId = "createMove")]
         public async Task<IActionResult> Create([FromBody] Move move, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<Evaluation>(move.EvaluationId, [SystemPermission.EditEvaluations], [EvaluationPermission.EditEvaluation], ct))
+                throw new ForbiddenException();
+
             move.CreatedBy = User.GetId();
             var createdMove = await _moveService.CreateAsync(move, ct);
             return CreatedAtAction(nameof(this.Get), new { id = createdMove.Id }, createdMove);
@@ -105,6 +114,9 @@ namespace Cite.Api.Controllers
         [SwaggerOperation(OperationId = "updateMove")]
         public async Task<IActionResult> Update([FromRoute] Guid id, [FromBody] Move move, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<Evaluation>(move.EvaluationId, [SystemPermission.EditEvaluations], [EvaluationPermission.EditEvaluation], ct))
+                throw new ForbiddenException();
+
             move.ModifiedBy = User.GetId();
             var updatedMove = await _moveService.UpdateAsync(id, move, ct);
             return Ok(updatedMove);
@@ -125,10 +137,12 @@ namespace Cite.Api.Controllers
         [SwaggerOperation(OperationId = "deleteMove")]
         public async Task<IActionResult> Delete(Guid id, CancellationToken ct)
         {
+            if (!await _authorizationService.AuthorizeAsync<Evaluation>(id, [SystemPermission.EditEvaluations], [EvaluationPermission.EditEvaluation], ct))
+                throw new ForbiddenException();
+
             await _moveService.DeleteAsync(id, ct);
             return NoContent();
         }
 
     }
 }
-
